Move main into own file
