+++ src/septentrio_gnss_driver/node/main.cpp
@@ -1,0 +1,56 @@
+// *****************************************************************************
+//
+// © Copyright 2020, Septentrio NV/SA.
+// All rights reserved.
+//
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions are met:
+//    1. Redistributions of source code must retain the above copyright
+//       notice, this list of conditions and the following disclaimer.
+//    2. Redistributions in binary form must reproduce the above copyright
+//       notice, this list of conditions and the following disclaimer in the
+//       documentation and/or other materials provided with the distribution.
+//    3. Neither the name of the copyright holder nor the names of its
+//       contributors may be used to endorse or promote products derived
+//       from this software without specific prior written permission.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+// ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+// LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+// CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+// SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+// INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+// CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+// ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+// POSSIBILITY OF SUCH DAMAGE.
+//
+// ****************************************************************************
+
+#include <septentrio_gnss_driver/node/rosaic_node.hpp>
+
+/**
+ * @file main.cpp
+ * @date 01/12/21
+ * @brief Main function of the ROSaic driver:
+ */
+
+int main(int argc, char** argv)
+{
+    ros::init(argc, argv, "septentrio_gnss");
+  	
+    // The info logging level seems to be default, hence we modify log level
+    // momentarily.. The following is the C++ version of
+    // rospy.init_node('my_ros_node', log_level=rospy.DEBUG)
+    if (ros::console::set_logger_level(
+            ROSCONSOLE_DEFAULT_NAME,
+            ros::console::levels::Debug)) // debug is lowest level, shows everything
+        ros::console::notifyLoggerLevelsChanged();
+
+    rosaic_node::ROSaicNode
+        rx_node; // This launches everything we need, in theory :)
+    ros::spin();
+    
+    return 0;
+}
