+++ src/septentrio_gnss_driver/node/rosaic_node.cpp
@@ -201,22 +201,3 @@ void rosaic_node::ROSaicNode::getROSParams()
     // To be implemented: RTCM, raw data settings, PPP, SBAS ...
     this->log(LogLevel::DEBUG ,"Finished getROSParams() method");
 };
-
-int main(int argc, char** argv)
-{
-    ros::init(argc, argv, "septentrio_gnss");
-  	
-    // The info logging level seems to be default, hence we modify log level
-    // momentarily.. The following is the C++ version of
-    // rospy.init_node('my_ros_node', log_level=rospy.DEBUG)
-    if (ros::console::set_logger_level(
-            ROSCONSOLE_DEFAULT_NAME,
-            ros::console::levels::Debug)) // debug is lowest level, shows everything
-        ros::console::notifyLoggerLevelsChanged();
-
-    rosaic_node::ROSaicNode
-        rx_node; // This launches everything we need, in theory :)
-    ros::spin();
-    
-    return 0;
-}
